requirements.txt
pytest>=8.0
anthropic>=0.80
python-dotenv>=1.0
feedparser>=6.0
requests>=2.31
beautifulsoup4>=4.12
